Dian-Yang-Dev/PipeDiagrams-main/PipeDiagrams-main/requirements.txt
pillow >= 11.3.0
